SERVER-24024 If there is no "local" database at startup in queryableBackupMode, do not panic.
At startup, the logic that checks for version-specific bugs that may need to be
worked around attempts to open the "local" database to see what version of
mongod was last used to access the data files. In queryableBackupMode, if there
is no local database, this can legally fail, at least for some storage engines.
This patch recovers from that failure, rather than shutting down the server.

This patch also updates the read_only and read_only_sharded test suites, to test
cases where there is no "local" database, as this is the standard
queryableBackupMode use case.


Former-commit-id: 68ee19bbd79cd33280cb01af38129249a32dcf7a