SERVER-24467 Make the shard implementation of 'moveChunk' joinable
This change makes the shard implementation of the 'moveChunk' command join
a currently running migration if that migration has exactly the same
arguments as the call. This allows retriability of this command by the
balancer.

In addition, it makes the moveChunk command report success if the
specified chunk is already located on the desired destination shard.


Former-commit-id: 3b7177d627a7a63b98b3717d982e323457267e12