SERVER-25676 Change name of replica set in create_index_gle.js to avoid bug.
Until a proper fix for SERVER-25676 is implemented, this change will prevent
the continuous build system from encountering it. This works because the
bug only manifests when a getLastError follows a call to createIndexes on
a sharded system where the shard replica set names appear before the
individual host names in the replica set in lexicographical order.


Former-commit-id: 784cc21ab3e7e44da724092178d56cec53ffab24