SERVER-26775: Cursors may be on a view namespace.
When a ClientCursorMonitor is timing out a cursor, its namespace may
be on a view. One case where this can happen is when the cursor was
on a collection, but when the ClientCursorMonitor gets to timing it
out, that collection was dropped and its namespace was recreated as
a view.


Former-commit-id: 549c3f4315fd95431a662e7a8eff880f00b1458d