SERVER-27240 Replace ConnectBG with poll
It's unsafe to close a socket from another thread.  Also, after
returning EINTR, the connect call converts to an async call.  And on
non-linux systems that requires a fallback to poll/select to handle
errors.

Because of that, let's just do the connect without the background thread
at all, starting off with poll.


Former-commit-id: a1baabeee5694aa8c4ffa1827233684d6c7fcc49