SERVER-28600 Slice `catalog/collection` cycles.
The `Collection` class is a big part of the circular
dependencies into `catalog`.  A simple vtable slice
is insufficient -- it was also necessary to alter the
constructors of some of the other types, to allow them
to properly self-initialize.

The `Collection` class's header also exposed a few static
and free functions that needed slicing as well.  These slices
were done using local `stdx::function` objects in the
`collection.cpp` TU and creating registration functions to
facilitate dependency injection.


Former-commit-id: 0609c4c9b296e0b036ee7eecf1eea39ebbf1b6a7