SERVER-1475 Make $type:'array' match outer arrays.
Prior to this change, the semantics of $type were that it
only matches documents with nested arrays, e.g.  {x: {$type:
'array'}} would match {x: [1, [2, 3]]} but not {x: [1, 2,
3]}. This is inconsistent with the matching semantics for
other query predicate operators, which apply both to the
individual array elements and the array as a whole.

The new behavior is that single-level arrays as well as
nested arrays will match {$type:'array'}. This is a breaking
change for applications which rely on the old behavior.


Former-commit-id: c6d0c0dd01c328d8ed1a5744485b29826b1fb2c9