SERVER-29210: Manage WiredTiger checkpoints in MongoDB.
Prior to this change, WiredTiger would be started with a configuration that
instructs WiredTiger to start its own checkpoint thread. The parameters
would trigger a checkpoint every 60 seconds or 2GB of data written to the
journal. This patch disables WiredTiger's checkpoint thread and moves the
responsibility to MongoDB.

The checkpoint thread now only checkpoints every 60 seconds (still
configurable with `--syncdelay`). The trigger on the amount of data written
to the journal was removed as an unnecessary mechanism to maintain parity
with.


Former-commit-id: cd2e5f07df13612b68deea94b82b001fca119373