SERVER-27244 Status usage compile-time facilities.
There are numerous places in the codebase where `mongo::Status` or
`mongo::StatusWith< T >` objects are returned and never checked.
Many of these are innocuous, but many of them are potentially
severe bugs.  This change introduces facilities to permit
compile-time warning of unchecked `Status` and `StatusWith` usage
on clang compilers.  It introduces an `ignore` function which is
useful to state that a specific "ignored status" case was
intentional.  It not presently an error, in clang builds, to
forget to check a `Status` -- this will come in a later commit.

This also introduces a `transitional_ignore` function, which allows
for easy continual auditing of the codebase for current "whitelisted"
unchecked-status instances.  All present "ignored status" cases
have been marked `transitional_ignore`.


Former-commit-id: 9abef6f25aadfd04309cb2219068097f93dc961d