SERVER-28765 Create BitNode.
Additionally, this change fixes a couple of minor things. First, it
makes the single-argument constructor for ArithmeticNode
explicit. Second, it adds a (noop = false) assignment in the else
branch of a test in set_node.cpp, instead of relying on the fact that
callers set noop to false. That gives SetNode clearer behavior.


Former-commit-id: 88c47189681b3212fe7ad98e4097b7e00b1aa232