SERVER-31355 Synchronize stepdown in auto_retry_on_network_error.js
Changes the test to wait for all client connections to be closed before
attempting to run commands on the replica set connection object.

Also introduces a failpoint to cause the server to hang after it has
updated its member state to SECONDARY and before it has begun to close
all of its client connections.


Former-commit-id: d4eb562ac63717904f24de4a22e395070687bc62