SERVER-30751 Use optimized MatchExpressions in JSON parser test.
The JSON Schema parser tests check that the output of the parser
matches an expected MatchExpression. The initial output of the parser
often includes a lot of unnecessary $and and $or stages that make it
confusing to read. Optimizing this output and using that for the check
instead makes the test more readable and adds some additional testing
for the optimizer.

This patch also adds getOptimizer() logic to a few MatchExpression
classes that still needed it.


Former-commit-id: c981f8578b30c3b8d86dc06605feb292264cebc2