SERVER-32233 Ensure electionId is increasing per iteration
NewPrimaryWithMaxElectionId test should ensure that electionId is newer
every iteration.

The problem is that the test assumes that OID generated per iteration is
increasing. And this assumption is based on the fact that OID::gen
constructs a value based on the return value from time(). However, there
is no guarantee that time() will consistently be returning values that
increases over time.

Use OID::fromTerm() in place of OID::gen().

Closes #1207

Signed-off-by: Randolph Tan <randolph@10gen.com>

Former-commit-id: 538354fddd70b3b2de8ea13f2f8dec72c973db87