SERVER-32822 Support read from unique index with both old and new format entries
During rolling upgrades the existing unique indexes' version would be changed
without translating the existing data to the new format. With an index in this
state, new data can be inserted in new format. When reading from such an index,
use the correct read logic based on the format of the entry being read.

In doing so, filter the key from the KeyString to obtain size of the Index Key.
Index Key size would indicate that whether the key is old format or new format.


Former-commit-id: 9960da7396adf3a92ac22e3c0fc62be1bc54be91