SERVER-33846: Set oplog visibility synchronously on secondaries, at the end of every batch.
This patch introduces an optimization to allow secondaries to set their
visibility synchronously with oplog application as well as bypassing
additional journal flushing meant for primaries. Primaries replicating oplog
entries atomically generate a new optime and pass it to the storage engine's
oplog record store via the `oplogDiskLocRegister` method. This code path
will now pass in a parameter `orderedCommit = false` that alerts the storage
engine to maintain the necessary oplog visibility semantics for that write.
This is existing behavior, the only difference is the addition of the
`orderedCommit` parameter.

Secondaries will now also call `oplogDiskLocRegister` at the end of every
batch. This call will pass in the optime of the last oplog entry applied and
`orderedCommit = true`. A storage engine may take this as a guarantee that
there are no oplog holes prior to the input optime.


Former-commit-id: 8ab8025f23fb50d15336870a6cea4e4ca6f9673a