SERVER-33909 More detailed error reporting from CAPI
The `libmongodbcapi.h` file now contains detailed documentation of
all preconditions, postconditions, and behaviors.  All operations
now take a `libmongodbcapi_status *` argument which will be
populated with the failure results of an operation.

The internals of how the `libmongodbcapi_` functions are implemented
have been rewritten to catch and report nearly all failures, as well
as to use more native C++ idioms such as exceptions and RAII.


Former-commit-id: 6b23bd026bb74b1e44d93951662070076e99b361